Creates a custom hook for auth operations
